new: allow job and job_signup cancellation and send email notifications
This became quite of a big digression while playing with the idea of
canceling jobs and job_signups instead of deleting them from the DB
(i.e. with :destroy). I learnt how to implement an 'includable' module to
gather this functionality. This module, called 'Cancelable', can be
included in ActiveRecord models providing the functionality at once. It
was also interesting how to implement 'before_' and 'after_' callbacks.
As an application of these callbacks, I implemented email notifications
when a job or a job_signup are canceled.