fix: temporarily publish letencrypt challenge via ENV variable
The previously created rake task only worked locally. When executing the
rake task on Heroku via:

  $ heroku run rake letsencrypt:publish_challenge[filename,challenge]

it ran on a separated dyno, whose filesystem was separated from the
running, and deleted when the dyno was terminated.

Now we create the files on application start if the corresponding ENV
variable LETSENCRYPT_CHALLENGE=filename,challenge is setup.