fix: tmp remove .erb tags in .coffee to try to fix deployment to heroku
When Travis-CI deploys to Heroku, the deployment fails while preparing app for
Rails asset pipeline and running 'rake assets:precompile'. The first
line of the error states:

 ActiveRecord::StatementInvalid: PG::UndefinedTable: ERROR:  relation
 "subscriptions" does not exist

Therefore, I think that because the file subscriptions.coffee.erb (edited in
this commit) made some references to the Subscription class, it failed
to load the class because the migrations where not yet applied.

If this fixes the deployment, we'll be able to revert this commit
because the migrations will be applied.