change: gracefully handle google maps for non-geocoded items
This showed up for the depots maps. Normally (in production) they
will always properly be geocoded, because they won't be valid otherwise.
However, while seedeing the DB with seed.rb, the addresses of the depots
are not validated, thus they are nor geocoded, thus the map cannot be
shown.

Previously the GoogleMap helpers where raising an exception if a item was
not geocoded. Now we display a placeholder alert informing about the
issue. However, again, this should only be seen during develoment.