fix: allow the subscription modification week to be equal to the current one
It happened that the app restarted on its own (expected behaviour) and
then the value of the SUBSCRIPTION_NEXT_UPDATE_WEEK_NUMBER was the same
as the current week, thus the value was ignored as it was expected to be
in an stric future week, and some users were not able to update the
subscription anymore close to the deadline.

I'm not adding a test for this because triky to test the value of an ENV
variable which depends on the current date.