fix: properly define config constant for handled error codes
Placing the constant in the ErrorsController class, and then using it in
the routes.rb file as ErrorsController::ERROR_CODES was causing to
load the ErrorsController, which in its turn loads its parent class
ApplicationController, which in its turn loads lots of helpers, causing
one of those helpers, PublicActivityHelper, to be loaded too early if
the DB schema was not already loaded, causing some tests to fail.

This should thus allow to clear the FIXME in the .travis.yml file added
with commit 0182b41 because 'db:schema:load' should not be necessary
anymore.