fix: remove erb preprocessing of coffee files
Although preprocessing coffee files with erb is possible, it adds
sometimes unexpected dependencies. For example, in the file
subscriptions.coffee (fixed in this commit) we were loading some
constant values from the Subscription model (with Subscription::*). This
apparently innofensive call rendered the application undeployable on a
new Heroku dyno. The problem was that during asset precompilation the whole
Subscription class had to be loaded, which implied also that all required
files in the Subscription class were also loaded. On of the files
(cancelable.rb) did a check on the database for a column, but the database
had not been migrated at that point yet (on a new Heroku dyno, that is)
causing the application to fail. As a general rule, it seems much safer
to pass the model values in the view as data-* attributes and load them
from the coffeescript file.